refactor(db): split articles.ts into cursor/query/write modules (#399)
- articles-cursor.ts: Cursor 型, appendCursorCondition, extractWithCursor,
  buildPaginatedQuery, ARTICLES_SELECT_FIELDS/FROM_JOIN 定数
- articles-query.ts: コア read 系 (list/related/neighbors/feed/author/category/day)
- articles-query-extra.ts: 月次 / calendar / searchArticles / countArticlesByDay
- articles-stats.ts: count / トレンド / stats 系 (countAll, trend30d, feedActivity 等)
- articles-write.ts: insertArticles / findExistingArticleGuids / buildInsertArticleStmts
- articles.ts: 薄い re-export ファイルとして維持 (既存 import を壊さない)

副次的改善:
- GetArticlesByFeedResult / ByAuthor / ByCategory / ByDay を CursorPage 型エイリアスに統一
- getArticlesByFeed/Author/Category/Day の cursor pagination 骨格を buildPaginatedQuery に集約
- articles-cursor.test.ts 新規追加: buildPaginatedQuery / appendCursorCondition /
  extractWithCursor の単体テスト (実 D1 ベース, AAA 構造, expect.soft 活用)
