+++ apps/web/worker/db/articles-query-extra.ts
@@ -1,0 +1,207 @@
+import type { Article, FeedCategory, FeedLang } from "../types";
+import {
+  ARTICLES_FROM_JOIN,
+  ARTICLES_SELECT_FIELDS,
+  type Cursor,
+  appendCursorCondition,
+  extractWithCursor,
+} from "./articles-cursor";
+
+// ---------------------------------------------------------------------------
+// 月次系
+// ---------------------------------------------------------------------------
+
+export interface GetArticlesByMonthParams {
+  category?: FeedCategory;
+  lang?: FeedLang;
+  limit?: number;
+}
+
+export async function getArticlesByMonth(
+  db: D1Database,
+  year: number,
+  month: number,
+  opts?: GetArticlesByMonthParams,
+): Promise<Article[]> {
+  const start = new Date(Date.UTC(year, month - 1, 1)).toISOString();
+  const end = new Date(Date.UTC(year, month, 1)).toISOString();
+  const limit = Math.min(Math.max(opts?.limit ?? 200, 1), 500);
+
+  const conds: string[] = [`a.published_at >= ?1`, `a.published_at < ?2`];
+  const binds: unknown[] = [start, end];
+
+  if (opts?.category) {
+    conds.push(`a.category = ?${binds.length + 1}`);
+    binds.push(opts.category);
+  }
+  if (opts?.lang) {
+    conds.push(`a.lang = ?${binds.length + 1}`);
+    binds.push(opts.lang);
+  }
+
+  const where = `WHERE ${conds.join(" AND ")}`;
+  const sql = `
+    SELECT ${ARTICLES_SELECT_FIELDS}
+    ${ARTICLES_FROM_JOIN}
+    ${where}
+    ORDER BY a.published_at DESC, a.id DESC
+    LIMIT ?${binds.length + 1}
+  `;
+  binds.push(limit);
+
+  const result = await db
+    .prepare(sql)
+    .bind(...binds)
+    .all<Article>();
+  return result.results ?? [];
+}
+
+export interface CountArticlesByMonthParams {
+  category?: FeedCategory;
+  lang?: FeedLang;
+}
+
+export async function countArticlesByMonth(
+  db: D1Database,
+  year: number,
+  month: number,
+  opts?: CountArticlesByMonthParams,
+): Promise<number> {
+  const start = new Date(Date.UTC(year, month - 1, 1)).toISOString();
+  const end = new Date(Date.UTC(year, month, 1)).toISOString();
+
+  const conds: string[] = [`published_at >= ?1`, `published_at < ?2`];
+  const binds: unknown[] = [start, end];
+
+  if (opts?.category) {
+    conds.push(`category = ?${binds.length + 1}`);
+    binds.push(opts.category);
+  }
+  if (opts?.lang) {
+    conds.push(`lang = ?${binds.length + 1}`);
+    binds.push(opts.lang);
+  }
+
+  const where = `WHERE ${conds.join(" AND ")}`;
+  const sql = `SELECT COUNT(*) AS c FROM articles ${where}`;
+
+  const row = await db
+    .prepare(sql)
+    .bind(...binds)
+    .first<{ c: number }>();
+
+  return row?.c ?? 0;
+}
+
+// ---------------------------------------------------------------------------
+// calendar
+// ---------------------------------------------------------------------------
+
+export interface CalendarItem {
+  date: string;
+  count: number;
+}
+
+/** 指定日数分の日別記事数を返す。0 件の日は含まない */
+export async function getArticlesCalendar(
+  db: D1Database,
+  days: number,
+  lang?: FeedLang | null,
+  category?: FeedCategory | null,
+): Promise<CalendarItem[]> {
+  const conds: string[] = [`published_at >= datetime('now', ?1)`];
+  const binds: unknown[] = [`-${days} days`];
+
+  if (lang) {
+    conds.push(`lang = ?${binds.length + 1}`);
+    binds.push(lang);
+  }
+  if (category) {
+    conds.push(`category = ?${binds.length + 1}`);
+    binds.push(category);
+  }
+
+  const where = `WHERE ${conds.join(" AND ")}`;
+  const sql = `
+    SELECT strftime('%Y-%m-%d', published_at) AS date, COUNT(*) AS count
+    FROM articles
+    ${where}
+    GROUP BY strftime('%Y-%m-%d', published_at)
+    ORDER BY date ASC
+  `;
+
+  const result = await db
+    .prepare(sql)
+    .bind(...binds)
+    .all<CalendarItem>();
+
+  return result.results ?? [];
+}
+
+// ---------------------------------------------------------------------------
+// searchArticles
+// ---------------------------------------------------------------------------
+
+export interface SearchArticlesResult {
+  articles: Article[];
+  nextCursor: Cursor | null;
+}
+
+/**
+ * LIKE ベースの全文検索。
+ * token ごとに LOWER(title || ' ' || COALESCE(summary, '')) LIKE '%token%' を AND で連鎖する。
+ * FTS5 への切替時はこの関数を差し替えるだけで済むようにクリーンに分離している。
+ */
+export async function searchArticles(
+  db: D1Database,
+  tokens: string[],
+  limit: number,
+  cursor: Cursor | null,
+): Promise<SearchArticlesResult> {
+  const conds: string[] = [];
+  const binds: unknown[] = [];
+
+  for (const token of tokens) {
+    // LIKE のメタ文字をエスケープしてプレースホルダ経由で渡す
+    const escaped = token.replace(/[%_\\]/g, (ch) => `\\${ch}`);
+    conds.push(
+      `LOWER(a.title || ' ' || COALESCE(a.summary, '')) LIKE ?${binds.length + 1} ESCAPE '\\'`,
+    );
+    binds.push(`%${escaped}%`);
+  }
+
+  appendCursorCondition(conds, binds, cursor);
+
+  const where = conds.length > 0 ? `WHERE ${conds.join(" AND ")}` : "";
+  const sql = `
+    SELECT ${ARTICLES_SELECT_FIELDS}
+    ${ARTICLES_FROM_JOIN}
+    ${where}
+    ORDER BY a.published_at DESC, a.id DESC
+    LIMIT ?${binds.length + 1}
+  `;
+  binds.push(limit + 1);
+
+  const result = await db
+    .prepare(sql)
+    .bind(...binds)
+    .all<Article>();
+  return extractWithCursor(result.results ?? [], limit);
+}
+
+// ---------------------------------------------------------------------------
+// countArticlesByDay
+// ---------------------------------------------------------------------------
+
+/** 指定日 (UTC) の総記事数を返す */
+export async function countArticlesByDay(
+  db: D1Database,
+  startIso: string,
+  endIso: string,
+): Promise<number> {
+  const row = await db
+    .prepare(`SELECT COUNT(*) AS c FROM articles WHERE published_at >= ?1 AND published_at < ?2`)
+    .bind(startIso, endIso)
+    .first<{ c: number }>();
+  return row?.c ?? 0;
+}
